+++ src/Ui5/AbstractUi5Resource.php
@@ -2,6 +2,7 @@
 
 namespace LaravelUi5\Core\Ui5;
 
+use LaravelUi5\Core\Enums\ArtifactType;
 use LaravelUi5\Core\Ui5\Contracts\Ui5ModuleInterface;
 use LaravelUi5\Core\Ui5\Contracts\Ui5ResourceInterface;
 
@@ -14,5 +15,10 @@ abstract class AbstractUi5Resource implements Ui5ResourceInterface
     public function getModule(): Ui5ModuleInterface
     {
         return $this->module;
+    }
+
+    public function getType(): ArtifactType
+    {
+        return ArtifactType::Resource;
     }
 }