+++ CHANGELOG.md
@@ -13,7 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Moved `getType` into abstract classes
 - Refactored directory layout for `Ui5DashboardInterface` instances
 - Refactored directory layout for `Ui5ReportInterface` instances
-- 
+- Refactored directory layout for `Ui5ResourceInterface` instances
+
 ## [3.4.2] - 2026-02-07
 - Allow public access to core lib